requirements_solana.txt
# x402 Insurance - Solana Requirements
# Python dependencies for Solana deployment

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Solana SDK
solana==0.32.0
solders>=0.20.0,<0.21.0
anchorpy==0.19.1

# Cryptography (Solana ed25519)
PyNaCl==1.5.0
base58==2.1.1

# HTTP Client
httpx>=0.23.0,<0.24.0
requests==2.31.0

# zkEngine (included in repo)
# No external dependency needed

# Database
psycopg2-binary==2.9.9  # Optional: PostgreSQL support

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2

# Logging
python-json-logger==2.0.7

# Development
pytest==7.4.3
pytest-asyncio==0.23.2
black==23.12.1
